requirements.txt
sed_eval==0.2.1
torchlibrosa=0.0.4
autoth==0.0.3